fixed mantis 0017344: Failed test: Inhalte zur Wiederholung zu einer Frage hinzufügen (KEINE ABBRUCH SCHALTFLÄCHE)

+++ Modules/TestQuestionPool/classes/class.assQuestionGUI.php
@@ -1148,6 +1148,11 @@ abstract class assQuestionGUI
 		$template->setVariable("TEXT_QUESTION_TYPE", assQuestion::_getQuestionTypeName($this->object->getQuestionType()));
 		return $template->get();
 	}
+	
+	public function showSuggestedSolution()
+	{
+		$this->suggestedsolution();
+	}
 
 	/**
 	* Allows to add suggested solutions for questions
@@ -1158,15 +1163,16 @@ abstract class assQuestionGUI
 	{
 		global $ilUser;
 		global $ilAccess;
-		
-		if ($_POST["deleteSuggestedSolution"] == 1)
+
+		$save = (is_array($_POST["cmd"]) && array_key_exists("suggestedsolution", $_POST["cmd"])) ? TRUE : FALSE;
+
+		if ($save && $_POST["deleteSuggestedSolution"] == 1)
 		{
 			$this->object->deleteSuggestedSolutions();
 			ilUtil::sendSuccess($this->lng->txt("msg_obj_modified"), true);
 			$this->ctrl->redirect($this, "suggestedsolution");
 		}
 
-		$save = (is_array($_POST["cmd"]) && array_key_exists("suggestedsolution", $_POST["cmd"])) ? TRUE : FALSE;
 		$output = "";
 		$solution_array = $this->object->getSuggestedSolution(0);
 		$options = array(
@@ -1303,7 +1309,11 @@ abstract class assQuestionGUI
 				$form->addItem($hidden);
 				$form->addItem($question);
 			}
-			if ($ilAccess->checkAccess("write", "", $_GET['ref_id']))	$form->addCommandButton("suggestedsolution", $this->lng->txt("save"));
+			if ($ilAccess->checkAccess("write", "", $_GET['ref_id']))
+			{
+				$form->addCommandButton('showSuggestedSolution', $this->lng->txt('cancel'));
+				$form->addCommandButton('suggestedsolution', $this->lng->txt('save'));
+			}
 			
 			if ($save)
 			{
@@ -1330,6 +1340,7 @@ abstract class assQuestionGUI
 					}
 				}
 			}
+			
 			$output = $form->getHTML();
 		}
